requirements.txt
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
SQLAlchemy>=2.0.36
pyyaml>=6.0.2
psycopg2-binary>=2.9.10